Move Incr & Scope into a new Data.Scope module.
